[labs/virtualizer] Export event classes through a new events.js (#3430)
* Added an events.js to export the RangeChangedEvent and VisibilityChangedEvent classes.

* Point to events.js to get the event classes now and moved the custom Range interface into there.

* Added events.js files to package.json's files property and the wireit outputs for build:ts.
